added some basic HTTP status code error handling
Apple's fulfillment API is experiencing an extended outage (10/09/22), and InventoryWatch doesn't handle this situation particularly gracefully, showing a message indicating that there could be something wrong with the user's settings. The app should be updated to indicate, in this situation, that the issue appears to lie with Apple's service, not InventoryWatch.

fixes #86, bumps app version number